loading data by type/group
the reality is you can only load data by type / group IF they do not reference any nodes in front of them - any invisible ones who are established later.
if you make references through both links and nodes to whoever in other groups you want to link with, then they are all shown if you loop through the inside array (with j)
if you take out the array in the links, then that is evaluated and detached as expected
in summary, makes more sense to write the story sequentially OR NOT sequentially - and add things in accordingly. if more than one level should be shown then that could be done. but does that make sense in the story. as in, more than one level LATER or higher in the list of ids.